📝 (aa-ebpf-common): Add lib.rs with module declarations and no_std doc
